requirements.txt
# torch and torchvision are installed separately with CUDA support
# see setup_and_train.bat
diffusers>=0.25.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0
numpy>=1.24.0
bitsandbytes>=0.41.0
wandb>=0.16.0
safetensors>=0.4.0
matplotlib>=3.7.0
lpips>=0.1.4
